Properly handle inbuf > 0 on EOF
Explicitly run the read callback if EOF is received while the input
buffer has still unprocessed data.  This prevents the loss of messages
that are transmitted over the connection just before closing it, such as
SMTP server goodbye messages in response to the QUIT command.
Previously, such messages were randomly received or lost.