Refactor to make automatic and fix variable naming
Remove the need for an explicit NOPERMCHANGE flag by checking the output
of `id -u` and removing the change ownership flags on install when
called as non-root user.

Use a more accurate wording using OWNER instead of PERM, and add FLAG
where appropriate to denote that the variables contain full flags, not
just values.

Issue:		#141